[reboot]: Stop mux before reboot on dual ToR (#1500)
If the script detects it is running on a dual ToR system, stop the mux container before shutting down syncd

This is necessary to prevent some links from flapping when the standby ToR is rebooted - when reboot stop syncd, the linkmgr will stop receiving heartbeats and will then tell the mux to switchover. However, since the device is rebooting, no heartbeats will be sent out. The peer ToR will see this absence of heartbeats and tell the mux to switch back to itself (the peer ToR), causing the flap.

By stopping the mux container before shutting down syncd, we prevent the rebooting ToR from switching the mux.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>