[202012]Fixing db_migrator for Feature table  (#1676)
This is a cherrypick of https://github.com/Azure/sonic-utilities/pull/1674 for 202012 since the master checks are failing and we need to unblock the 202012 branch

#### What I did
Fixing db_migrator.py for FEATURE table. In case of version_unknown, all versions' migrator APIs would be invoked which will also invoke the feature_table migration. Without the field check for 'status' the current logic will set it 'disabled' and will set 'state' field too to be 'disabled' which will in turn disable all the features bringing down the system.

#### How I did it
Added check to migrate only if 'status' field is present.

#### How to verify it
Install the image through onie and confirm if services are up