Disable pfcwd forward action config for cisco-8000. (#1848)
Disable pfcwd forward action config for cisco-8000.
Do not configure pfcwd if pfc is not configured on a port
This PR is in association with: Azure/sonic-swss#1748 where PFC-WD forward action is disabled for cisco-8000 platform.

Signed-off-by: Alpesh S Patel alpesh@cisco.com

What I did
There are two changes of interest:
- disables pfcwd forward action configuration from pfcwd command
- in the pfcwd script, adds a check to enable pfcwd configuration on a port only if it has pfc enabled. If config_db.json has an empty entry like: "pfc_enable" : "", and subsequently pfcwd is configured, swss/orchagent throws an error logs

How to verify it
- can be verified on cisco-8000 platform. It is a no-op for other platforms
# pfcwd start -a forward 400 -r 400
SKIPPED: PFC WD 'forward' action not supported on asic 'cisco-8000'

- for a single port, in /etc/sonic/config_db.json, set no pfc enable bits by setting:
"pfc_enable" :  "" for atleast 1 port.
without this change, see the logs in syslog. With this change, there will just be a log on console.