Added cisco show platform commands (#2221)
What I did
Gave cisco-8000 the ability to add sub-commands under show platform <> in our downstream repo. Currently every time we want to add/remove/update a cli, we must raise a PR upstream.

How I did it
I have the cisco-8000.py module import a list of click commands that are written in a module that is located in our platform code.

How to verify it
Run show platform -h to see all commands. We will be able to see show platform inventory. This is only available on cisco devices.